Refactor src/tui/App.tsx into focused hooks
App.tsx grew to 1001 lines; split it into orchestration shell
(~310 lines) plus six focused hooks, one submit module, and a
TabPanels component. Behavior preserved end-to-end.

- src/tui/messages.ts — msgId, FOOTER_RESERVE
- src/tui/keys.ts — TAB_BY_CTRL_KEY
- src/tui/hooks/useTerminalRows — stdout rows + resize listener
- src/tui/hooks/useChatSession — session lifecycle, splash, performShutdown
- src/tui/hooks/useMessageQueue — queue + streaming engine
- src/tui/hooks/useChatTitlePolling — 5s title refresh
- src/tui/hooks/useCaptureTabCycle — env-driven tab cycling
- src/tui/hooks/useAppKeybindings — global keybindings
- src/tui/handleSubmit.ts — useChatSubmit incl /clear IIFE
- src/tui/components/TabPanels.tsx — tabs 2-8
