+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "botholomew",
-  "version": "0.16.2",
+  "version": "0.16.3",
   "description": "An autonomous AI agent for knowledge work — works your task queue while you sleep.",
   "type": "module",
   "bin": {
